Create daily notifications to be sent
Builder gets billable subscriptions of the day, a Notifier.Notification is created aswell as a SubNotification with the billable subscriptions.

Notifier.Notification does not care about Subs at all, it does not hold user or subscriptions.

Subs creates a SubsNotification which knows about the user and subscription and has a Notifier.Notification associated.

Given that Notifier and Subs have their own repo, we cannot work with transactions so there's a test failing. I'm going to extract Subs.Repo to a separated app in order to be shared by these two apps.